backend/requirements.txt
Flask~=3.0.3
flask-cors~=5.0.0
pillow~=10.4.0
pymongo~=4.3.3
python-dotenv~=0.21.0

